Splitting integration tests and tidying them up
