+++ api-search/src/integrationTest/groovy/org/cedar/onestop/api/search/CollectionSearchIntegrationTests.groovy
@@ -1,5 +1,6 @@
 package org.cedar.onestop.api.search
 
+import groovy.json.JsonOutput
 import org.apache.http.HttpEntity
 import org.apache.http.entity.ContentType
 import org.apache.http.nio.entity.NStringEntity
@@ -9,7 +10,9 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.web.server.LocalServerPort
+import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
+import org.springframework.http.RequestEntity
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.web.client.RestTemplate
 import spock.lang.Specification
@@ -24,6 +27,9 @@ class CollectionSearchIntegrationTests extends Specification {
 
   @Autowired
   RestClient restClient
+
+  @Autowired
+  Map testData
 
   @LocalServerPort
   private String port
@@ -47,12 +53,7 @@ class CollectionSearchIntegrationTests extends Specification {
   void setup() {
     def cl = ClassLoader.systemClassLoader
     def searchCollectionIndexJson = cl.getResourceAsStream('search_collectionIndex.json').text
-    def searchGranuleIndexJson = cl.getResourceAsStream('search_granuleIndex.json').text
-    def searchFlattenedGranuleIndexJson = cl.getResourceAsStream('search_flattened_granuleIndex.json').text
-
     def collectionIndexSettings = new NStringEntity(searchCollectionIndexJson, ContentType.APPLICATION_JSON)
-    def granuleIndexSettings = new NStringEntity(searchGranuleIndexJson, ContentType.APPLICATION_JSON)
-    def flattenedGranuleIndexSettings = new NStringEntity(searchFlattenedGranuleIndexJson, ContentType.APPLICATION_JSON)
 
     Response response = restClient.performRequest('DELETE', '_all')
     println("DELETE _all: ${response}")
@@ -60,63 +61,7 @@ class CollectionSearchIntegrationTests extends Specification {
     def collectionResponse = restClient.performRequest('PUT', COLLECTION_SEARCH_INDEX, Collections.EMPTY_MAP, collectionIndexSettings)
     println("PUT new collection index: ${collectionResponse}")
 
-    def granuleResponse = restClient.performRequest('PUT', GRANULE_SEARCH_INDEX, Collections.EMPTY_MAP, granuleIndexSettings)
-    println("PUT new granule index: ${granuleResponse}")
-
-    def flattenedGranuleResponse = restClient.performRequest('PUT', FLATTENED_GRANULE_SEARCH_INDEX, Collections.EMPTY_MAP, flattenedGranuleIndexSettings)
-    println("PUT new flattened-granule index: ${flattenedGranuleResponse}")
-
-    Map data = [
-        'DEM': [
-            'C1': [
-                id: 'e7a36e60-1bcb-47b1-ac0d-3c2a2a743f9b',
-                granules: [],
-                flattenedGranules: []
-            ],
-            'C2': [
-                id: 'e5820283-3686-44d0-8edd-28a086eb500e',
-                granules: [],
-                flattenedGranules: []
-            ],
-            'C3': [
-                id: '1415b3db-c602-4dbb-a502-4091fe9df1cf',
-                granules: [],
-                flattenedGranules: []
-            ],
-        ],
-        'GHRSST': [
-            'C1': [
-                id: '920d8155-f764-4777-b7e5-14442b7275b8',
-                granules: [],
-                flattenedGranules: []
-            ],
-            'C2': [
-                id: '882511bc-e99e-4597-b634-47a59ddf9fda',
-                granules: [],
-                flattenedGranules: []
-            ],
-            'C3': [
-                id: '42ea683d-e4e7-434c-8823-abff32e00f34',
-                granules: [],
-                flattenedGranules: []
-            ],
-        ],
-        'COOPS': [
-            'C1': [
-                id: 'fcf83ec9-964b-45b9-befe-378ea6ce52cb',
-                granules: [
-                    'G1': [id: '783089c4-3484-4f70-ac8d-d4818d0cd0dd'],
-                    'G2': [id: 'a207b48f-29fc-4d79-a676-1f265cd9971f'],
-                ],
-                flattenedGranules: [
-                    'FG1': [id: '783089c4-3484-4f70-ac8d-d4818d0cd0dd'],
-                    'FG2': [id: 'a207b48f-29fc-4d79-a676-1f265cd9971f']
-                ]
-            ]
-        ]
-    ]
-
-    data.each{ name, dataset ->
+    testData.each{ name, dataset ->
       dataset.each { collection, collectionData ->
         def metadata = cl.getResourceAsStream("data/${name}/${collection}.json").text
         def id = collectionData.id
@@ -124,26 +69,10 @@ class CollectionSearchIntegrationTests extends Specification {
         HttpEntity record = new NStringEntity(metadata, ContentType.APPLICATION_JSON)
         response = restClient.performRequest('PUT', collectionEndpoint, Collections.EMPTY_MAP, record)
         println("PUT new collection: ${response}")
-
-        collectionData.granules.each { granule, granuleData ->
-          metadata = cl.getResourceAsStream("data/${name}/${granule}.json").text
-          def granuleEndpoint = "/$GRANULE_SEARCH_INDEX/$TYPE/$granuleData.id"
-          record = new NStringEntity(metadata, ContentType.APPLICATION_JSON)
-          response = restClient.performRequest('PUT', granuleEndpoint, Collections.EMPTY_MAP, record)
-          println("PUT new granule: ${response}")
-        }
-
-        collectionData.flattenedGranules.each { flattenedGranule, flattenedGranuleData ->
-          metadata = cl.getResourceAsStream("data/${name}/${flattenedGranule}.json").text
-          def flattenedGranuleEndpoint = "/$FLATTENED_GRANULE_SEARCH_INDEX/$TYPE/$flattenedGranuleData.id"
-          record = new NStringEntity(metadata, ContentType.APPLICATION_JSON)
-          response = restClient.performRequest('PUT', flattenedGranuleEndpoint, Collections.EMPTY_MAP, record)
-          println("PUT new flattened granule: ${response}")
-        }
       }
     }
 
-    def refreshEndpoint = "/${COLLECTION_SEARCH_INDEX},${GRANULE_SEARCH_INDEX},${FLATTENED_GRANULE_SEARCH_INDEX}/_refresh"
+    def refreshEndpoint = "/${COLLECTION_SEARCH_INDEX}/_refresh"
     response = restClient.performRequest('POST', refreshEndpoint)
     println("Refresh all search indices: ${response}")
 
@@ -151,10 +80,338 @@ class CollectionSearchIntegrationTests extends Specification {
     restTemplate.errorHandler = new TestResponseErrorHandler()
     searchBaseUriString = "http://localhost:${port}/${contextPath}/search/"
     searchCollectionUriString = "http://localhost:${port}/${contextPath}/search/collection"
-    searchGranuleUriString = "http://localhost:${port}/${contextPath}/search/granule"
-    searchFlattenedGranuleUriString = "http://localhost:${port}/${contextPath}/search/flattened-granule"
     collectionBaseUriString = "http://localhost:${port}/${contextPath}/collection/"
-    granuleBaseUriString = "http://localhost:${port}/${contextPath}/granule/"
-    flatGranuleBaseUriString = "http://localhost:${port}/${contextPath}/flattened-granule/"
+  }
+
+  def 'Collection endpoint reports count of collections'() {
+    setup:
+    def searchCollectionBaseUri = (collectionBaseUriString).toURI()
+    def requestEntityCollection = RequestEntity.get(searchCollectionBaseUri).build()
+
+    when:
+    def resultCollection = restTemplate.exchange(requestEntityCollection, Map)
+
+    then:
+    resultCollection.statusCode == HttpStatus.OK
+    resultCollection.headers.getContentType() == contentType
+    resultCollection.body.data[0].count == 7
+    resultCollection.body.data*.id.containsAll(['collection'])
+    resultCollection.body.data*.count.every({ it instanceof Number })
+  }
+
+  def 'Valid query-only collection search summary returns OK with expected results'() {
+    setup:
+    def searchBaseUri = (searchCollectionUriString).toURI()
+    def request = """\
+        {
+          "queries":
+            [
+              { "type": "queryText", "value": "temperature"}
+            ]
+        }""".stripIndent()
+
+    def requestEntity = RequestEntity
+        .post(searchBaseUri)
+        .contentType(contentType)
+        .body(request)
+
+    when:
+    def result = restTemplate.exchange(requestEntity, Map)
+
+    then: "Search returns OK"
+    result.statusCode == HttpStatus.OK
+    result.headers.getContentType() == contentType
+
+    and: "Result contains 4 items"
+    def items = result.body.data
+    items.size() == 4
+
+    and: "Expected results are returned"
+    // Not returning IDs so need to check another way
+    def thumbnails = items.collect { it.attributes.thumbnail }
+    thumbnails.sort().equals([
+        'http://data.nodc.noaa.gov/cgi-bin/gfx?id=gov.noaa.nodc:GHRSST-Geo_Polar_Blended_Night-OSPO-L4-GLOB',
+        'http://data.nodc.noaa.gov/cgi-bin/gfx?id=gov.noaa.nodc:NDBC-COOPS',
+        'http://data.nodc.noaa.gov/cgi-bin/gfx?id=gov.noaa.nodc:GHRSST-EUR-L4UHFnd-MED',
+        'http://data.nodc.noaa.gov/cgi-bin/gfx?id=gov.noaa.nodc:GHRSST-OSDPD-L2P-MTSAT1R'
+    ].sort())
+
+    // If there are more keys per result than summary should contain, summary source filter is not working!
+    items.each {
+      assert it.attributes.keySet().size() <= [
+          "title",
+          "thumbnail",
+          "spatialBounding",
+          "beginDate",
+          "beginYear",
+          "endDate",
+          "endYear",
+          "links",
+          "citeAsStatements",
+          "internalParentIdentifier"
+      ].size()
+    }
+  }
+
+  def 'Valid query-only collection search with facets returns OK with expected results'() {
+    setup:
+    def searchBaseUri = (searchCollectionUriString).toURI()
+    def request = """\
+        {
+          "queries":
+            [
+              { "type": "queryText", "value": "temperature"}
+            ],
+          "facets" : true,
+          "summary" : false
+        }""".stripIndent()
+
+    def requestEntity = RequestEntity
+        .post(searchBaseUri)
+        .contentType(contentType)
+        .body(request)
+
+    when:
+    def result = restTemplate.exchange(requestEntity, Map)
+
+    then: "Search returns OK"
+    result.statusCode == HttpStatus.OK
+    result.headers.getContentType() == contentType
+
+    and: "Result contains 4 items"
+    def items = result.body.data
+    items.size() == 4
+
+    and: "Expected results are returned"
+    def actualIds = items.collect { it.attributes.fileIdentifier }
+    actualIds.containsAll([
+        'gov.noaa.nodc:GHRSST-EUR-L4UHFnd-MED',
+        'gov.noaa.nodc:GHRSST-OSDPD-L2P-MTSAT1R',
+        'gov.noaa.nodc:GHRSST-Geo_Polar_Blended_Night-OSPO-L4-GLOB',
+        'gov.noaa.nodc:NDBC-COOPS'
+    ])
+
+    and: 'The correct number of facets is returned'
+    def aggs = result.body.meta.facets
+    aggs.size() == 10
+
+    and: 'The facets are as expected'
+    aggs.science != null
+    aggs.services != null
+    aggs.locations != null
+    aggs.instruments != null
+    aggs.platforms != null
+    aggs.projects != null
+    aggs.dataCenters != null
+    aggs.horizontalResolution != null
+    aggs.verticalResolution != null
+    aggs.temporalResolution != null
+
+    and: 'The cleaned aggregations are actually cleaned'
+    def locationTerms = aggs.locations.collect { it }
+    // Bad planted keywords should be removed
+    !locationTerms.contains('Alaska')
+    !locationTerms.contains('Alaska > Unalaska')
+  }
+
+  def 'Valid filter-only collection search returns OK with expected results'() {
+    setup:
+    def searchBaseUri = (searchCollectionUriString).toURI()
+    def request = """\
+        {
+          "filters":
+            [
+              {"type": "geometry", "relation": "contains", "geometry": {"type": "Point", "coordinates": [145.5, 12.34]}}
+            ],
+          "facets": false,
+          "summary": false
+        }""".stripIndent()
+
+    def requestEntity = RequestEntity
+        .post(searchBaseUri)
+        .contentType(contentType)
+        .body(request)
+
+    when:
+    def result = restTemplate.exchange(requestEntity, Map)
+
+    then: "Search returns OK"
+    result.statusCode == HttpStatus.OK
+    result.headers.getContentType() == contentType
+
+    and: "Result contains 4 items"
+    def items = result.body.data
+    items.size() == 4
+
+    and: "Expected results are returned"
+    def actualIds = items.collect { it.attributes.fileIdentifier }
+    actualIds.containsAll([
+        'gov.noaa.nodc:GHRSST-Geo_Polar_Blended_Night-OSPO-L4-GLOB',
+        'gov.noaa.ngdc.mgg.dem:4870',
+        'gov.noaa.nodc:GHRSST-OSDPD-L2P-MTSAT1R',
+        'gov.noaa.nodc:NDBC-COOPS'
+    ])
+  }
+
+  def 'Valid query-and-filter collection search returns OK with expected result'() {
+    setup:
+    def searchBaseUri = (searchCollectionUriString).toURI()
+    def request = """\
+        {
+          "queries":
+            [
+              { "type": "queryText", "value": "temperature"}
+            ],
+          "filters":
+            [
+              {"type": "datetime", "before": "2007-12-31T23:59:59.999Z", "after": "2007-01-01T00:00:00Z"}
+            ],
+          "summary": false
+        }""".stripIndent()
+
+    def requestEntity = RequestEntity
+        .post(searchBaseUri)
+        .contentType(contentType)
+        .body(request)
+
+    when:
+    def result = restTemplate.exchange(requestEntity, Map)
+
+    then: "Search returns OK"
+    result.statusCode == HttpStatus.OK
+    result.headers.getContentType() == contentType
+
+    and: "Result contains 1 item"
+    def items = result.body.data
+    items.size() == 1
+
+    and: "Expected result is returned"
+    def actualIds = items.collect { it.attributes.fileIdentifier }
+    actualIds.containsAll([
+        'gov.noaa.nodc:GHRSST-EUR-L4UHFnd-MED'
+    ])
+  }
+
+  def 'Valid query-and-exclude-global collection search returns OK with expected results'() {
+    setup:
+    def searchBaseUri = (searchCollectionUriString).toURI()
+    def request = """\
+        {
+          "queries":
+            [
+              {"type": "queryText", "value": "ghrsst"}
+            ],
+          "filters":
+            [
+              {"type": "excludeGlobal", "value": true}
+            ],
+          "summary": false
+        }""".stripIndent()
+
+    def requestEntity = RequestEntity
+        .post(searchBaseUri)
+        .contentType(contentType)
+        .body(request)
+
+    when:
+    def result = restTemplate.exchange(requestEntity, Map)
+
+    then: "Search returns OK"
+    result.statusCode == HttpStatus.OK
+    result.headers.getContentType() == contentType
+
+    and: "Result contains 2 items"
+    def items = result.body.data
+    items.size() == 2
+
+    and: "Expected result is returned"
+    def actualIds = items.collect { it.attributes.fileIdentifier }
+    actualIds.containsAll([
+        'gov.noaa.nodc:GHRSST-EUR-L4UHFnd-MED',
+        'gov.noaa.nodc:GHRSST-OSDPD-L2P-MTSAT1R'
+    ])
+  }
+
+  def 'Time filter with #situation an item\'s time range returns the correct results'() {
+    setup:
+    def searchBaseUri = (searchCollectionUriString).toURI()
+    def ghrsst1FileId = 'gov.noaa.nodc:GHRSST-EUR-L4UHFnd-MED'
+    def request = [filters: [[type: 'datetime']]]
+    if (before) {
+      request.filters[0].before = before
+    }
+    if (after) {
+      request.filters[0].after = after
+    }
+
+    request.summary = false
+
+    def requestEntity = RequestEntity
+        .post(searchBaseUri)
+        .contentType(contentType)
+        .body(JsonOutput.toJson(request))
+
+    when:
+    def result = restTemplate.exchange(requestEntity, Map)
+    def ids = result.body.data.collect { it.attributes.fileIdentifier }
+
+    then:
+    result.statusCode == HttpStatus.OK
+    ids.contains(ghrsst1FileId) == matches
+
+    where: // NOTE: time range for GHRSST/1.xml is: 2005-01-30T00:00:00.000Z <-> 2008-01-14T00:00:00.000Z
+    after                  | before                 | matches | situation
+    '2005-01-01T00:00:00Z' | '2005-01-02T00:00:00Z' | false   | 'range that is fully before'
+    '2005-01-01T00:00:00Z' | '2008-01-01T00:00:00Z' | true    | 'range that overlaps the beginning of'
+    '2005-02-01T00:00:00Z' | '2008-01-01T00:00:00Z' | true    | 'range that is fully within'
+    '2005-01-01T00:00:00Z' | '2008-02-01T00:00:00Z' | true    | 'range that fully encloses'
+    '2005-02-01T00:00:00Z' | '2008-02-01T00:00:00Z' | true    | 'range that overlaps the end of'
+    '2008-02-01T00:00:00Z' | '2008-02-02T00:00:00Z' | false   | 'range that is fully after'
+    '2005-01-01T00:00:00Z' | null                   | true    | 'start time before'
+    '2005-02-01T00:00:00Z' | null                   | true    | 'start time within'
+    '2008-02-01T00:00:00Z' | null                   | false   | 'start time after'
+    null                   | '2005-01-01T00:00:00Z' | false   | 'end time before'
+    null                   | '2008-01-01T00:00:00Z' | true    | 'end time within'
+    null                   | '2008-02-01T00:00:00Z' | true    | 'end time after'
+  }
+
+  def 'Search with pagination specified returns OK with expected results'() {
+    setup:
+    def searchBaseUri = (searchCollectionUriString).toURI()
+    def request = """\
+        {
+          "queries":
+            [
+              { "type": "queryText", "value": "temperature"}
+            ],
+            "page":
+            {
+              "max": 1,
+              "offset": 0
+            },
+          "summary": false
+        }""".stripIndent()
+
+    def requestEntity = RequestEntity
+        .post(searchBaseUri)
+        .contentType(contentType)
+        .body(request)
+
+    when:
+    def result = restTemplate.exchange(requestEntity, Map)
+
+    then: "Search returns OK"
+    result.statusCode == HttpStatus.OK
+    result.headers.getContentType() == contentType
+
+    and: "Result contains 1 item"
+    def items = result.body.data
+    items.size() == 1
+
+    and: "Expected result is returned"
+    def actualIds = items.collect { it.attributes.fileIdentifier }
+    actualIds.containsAll([
+        'gov.noaa.nodc:GHRSST-Geo_Polar_Blended_Night-OSPO-L4-GLOB'
+    ])
   }
 }
